tail: Use getc and putc instead of concat
The FILE streams are buffered, so the end result is essentially the
same, except now lines are printed as they are read since we don't wait
until the entire buffer is filled.

This also fixes an issue where only part of the lines during the follow
were printed if they contained NUL bytes.