Don't recurse again if we are at the top level
recurse() runs fn for every file contained in the named directory, and
then one last time on the named directory if depth == 0 (to assist with
starting recurse). So, if depth == 0, don't recurse again. This doesn't
cause any issues since recurse checks against its history to prevent loops,
but we should skip the history lookup if we don't need it.